Cambios realizados

+++ web-renta-autos/src/main/java/com/renta/autos/models/service/implementations/AutomovilService.java
@@ -9,8 +9,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.renta.autos.models.entities.Accesorios;
 import com.renta.autos.models.entities.Automovil;
+import com.renta.autos.models.entities.Mantenimiento;
 import com.renta.autos.models.repositories.IAccesorios;
 import com.renta.autos.models.repositories.IAutomovil;
+import com.renta.autos.models.repositories.IMantenimiento;
 import com.renta.autos.models.service.interfaces.IAutomovilService;
 
 @Service
@@ -21,6 +23,9 @@ public class AutomovilService implements IAutomovilService{
 	
 	@Autowired
 	IAccesorios repositoryAccesorios;
+	
+	@Autowired
+	IMantenimiento repositoryMantenimiento;
 	
 	@Override
 	@Transactional
@@ -42,6 +47,16 @@ public class AutomovilService implements IAutomovilService{
 	@Override
 	@Transactional
 	public void delete(Integer codigo) {
+		Optional<Automovil> automovil = repository.findById(codigo);
+		if(automovil == null) return;
+		for(Accesorios accesorios : automovil.get().getListaAccesorios()) {
+			repositoryAccesorios.deleteById(accesorios.getCodigoAccesorios());
+		}
+		
+		for(Mantenimiento mantenimiento: automovil.get().getMantenimiento()) {
+			repositoryMantenimiento.deleteById(mantenimiento.getCodigoMantenimiento());
+		}
+		
 		repository.deleteById(codigo);
 	}
 
@@ -49,5 +64,11 @@ public class AutomovilService implements IAutomovilService{
 	@Transactional(readOnly = true)
 	public List<Automovil> findAll() {
 		return (List<Automovil>) repository.findAll();
+	}
+
+	@Override
+	@Transactional(readOnly = true)
+	public List<Automovil> findByDisponibilidad(boolean disponibilidad) {
+		return (List<Automovil>) repository.findByDisponibilidad(disponibilidad);
 	}
 }
